requirements.txt
requests
flask
pytest
pytest-cov
pyYaml
httmock